requirements.txt
torch
diffusers
transformers
accelerate
peft
numpy
Pillow
tqdm
